Add missing files for the darwin fix.
